+++ spec/support/namespace_factory.rb
@@ -2,7 +2,7 @@ module ActiveRecordTemporalTests
   module NamespaceFactory
     def history_model_namespace(name = "History", &block)
       mod = Module.new do
-        include SystemVersioning::HistoryModelNamespace
+        include ActiveRecord::Temporal::SystemVersioning::HistoryModelNamespace
 
         instance_eval(&block) if block
       end